requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
beautifulsoup4==4.9.1
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.3
djangorestframework==3.11.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
psycopg2==2.8.6
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-decouple==3.3
pytz==2020.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
win-inet-pton==1.1.0
wincertstore==0.2
